fix: color extension not working
