+++ packages/editor/src/core/extensions/index.ts
@@ -10,9 +10,8 @@ export * from "./slash-commands";
 export * from "./table";
 export * from "./typography";
 export * from "./core-without-props";
-export * from "./custom-background-color";
 export * from "./custom-code-inline";
-export * from "./custom-text-color";
+export * from "./custom-color";
 export * from "./drop";
 export * from "./enter-key-extension";
 export * from "./extensions";
