+++ packages/editor/src/core/extensions/read-only-extensions.tsx
@@ -21,8 +21,7 @@ import {
   CustomMention,
   HeadingListExtension,
   CustomReadOnlyImageExtension,
-  CustomTextColorExtension,
-  CustomBackgroundColorExtension,
+  CustomColorExtension,
 } from "@/extensions";
 // helpers
 import { isValidHttpUrl } from "@/helpers/common";
@@ -123,8 +122,7 @@ export const CoreReadOnlyEditorExtensions = (props: Props) => {
       readonly: true,
     }),
     CharacterCount,
-    CustomTextColorExtension,
-    CustomBackgroundColorExtension,
+    CustomColorExtension,
     HeadingListExtension,
   ];
 };
